requirements.txt
faiss_cpu==1.7.4
fastapi==0.104.1
numpy==1.23.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==10.1.0
Requests==2.31.0
torch==2.0.1+cu117
transformers==4.34.0
